Fix synchronous returns on blob.
Sometimes even the remote blobvault backend can callback synchronously - seems
to be the case in IE9 using XDomainRequest, perhaps due to caching. So we wrap
BlobObj.get in setImmediate to ensure it is always asynchronous, no matter which
backends or even if the backends are buggy.

It would be even cleaner to wrap the blob stuff in an Angular service, so we
should do that eventually.