Change OfferCancel transactions so they show up as a primary effect.
To cancel an offer is the primary effect of an OfferCancel transaction. So it
should be returned as such by the jsonrewriter.

There could still be a type of transaction in the future (e.g. a contract
transaction) that has an offer_cancelled side effect, so I left that code in.

As for the spelling of "cancelled", see:
http://grammarist.com/spelling/cancel/