For write access, we need to use the app.id instance still. Fixes #481.
The problem was that $id inherits from app.id, but any properties set on it are
local to it. Since app.id was still managing blob updates, blob updates did not
get promoted.

The bug did not manifest when the login happened via persistent_auth, which is
why our developers didn't notice it at first.