Usability: Wallet backups should be encrypted with case-insensitive username.
Usernames are usually treated case-insensitive and users have come to expect
that behavior. It was regression that caused backups to be created with a
case-sensitive username.

To fix, we'll ensure new backups are created case-insensitive again and when we
implement wallet restore-from-backup, we'll make sure to try both versions:
as-entered and lowercased username.