Switch to an Angular directive for showing QR codes.
Note that the underlying library is a little bit more low-level than the one we
were using before. For example it requires you to specify the QR version (number
between 1-40 that determines capacity) manually instead of auto-detecting it.

So in the future we might switch out the underlying library, while keeping the
rp-qrcode directive.