Support for federated login
Login information is already detected from the configured domain's
ripple.txt file.  This patch changes things so that if a username is
entered in "email address style" the auth information is detected from
the given domain instead of the locally-configured one.  This allows
users to easily choose alternative blobvaults without sacrificing client
support.