[FIX] Better error handling for JSON rewriter issues.
It would be nice if issues like RT-2341 didn't take down the entire history page, but rather showed the specific transaction(s) that failed to process. This patch adds the error handling necessary.

It also adds a new mixpanel event for JsonRewriter exceptions so that we find out if there are any transaction failing to parse out there.