[FIX] Send: fix currencies constraints (RT-3393)
Fix processing of currencies constraints.
When RT is loaded in browser, underscore lib gets replaced
by the lodash by something. And in the old versions of lodash
there was bug in _.intersection method (if argument is just one
array it returned this array). But in recent versions they fixed
this bug, so $scope.update_currency_choices broke.